test privatelink in stream connection
